Surfaced captured pip output in PythonInstaller failures
* Promoted TeeTextWriter from JsonUiService private inner class to public Util/UtilIO.cs
* Added RunProcessOrThrow helper that tees Writer + capture and throws with output on non-zero exit
* Refactored 5 RunProcess call sites in PythonInstaller to use the helper
* Added new resource string for Failed-to-execute-command-with-output

See ai/todos/active/TODO-20260430_pip_bootstrap_py39.md
