+++ pwiz_tools/Skyline/Model/Tools/ToolsResources.resx
@@ -265,6 +265,12 @@
   <data name="PythonInstaller_Failed_to_execute_command____0__" xml:space="preserve">
     <value>Failed to execute command: [{0}]</value>
   </data>
+  <data name="PythonInstaller_Failed_to_execute_command____0____Output____1__" xml:space="preserve">
+    <value>Failed to execute command: [{0}]
+
+Output:
+{1}</value>
+  </data>
   <data name="PythonInstaller__0__Running_command____1____2__" xml:space="preserve">
     <value>{0} Running command: [{1}] {2} </value>
   </data>
